Rename/#76: 온보딩 내부 페이지 컴포넌트 재명명
