+++ src/pages/intro/Intro.tsx
@@ -1,10 +1,10 @@
-import { Main, Second, Third, Forth } from './components';
+import { Main, AcrossSlide, PopList, ScrollPage } from './components';
 import * as S from './intro.styled';
 
 const Intro = () => {
   return (
     <S.Container>
-      {[Main, Second, Third, Forth].map((Component, index) => (
+      {[Main, AcrossSlide, PopList, ScrollPage].map((Component, index) => (
         <S.Section key={index}>
           <Component />
         </S.Section>
