+++ src/pages/intro/components/acrossSlide/AcorssSlide.styled.ts
