+++ src/pages/intro/components/popList/PopList.styled.ts
